# meet-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Api for meet application.",
        "title": "meet API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "paths": {
        "/v1/buy-pack/capture-payment": {
            "post": {
                "description": "Capture a Payment to a Pack",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BuyPack"
                ],
                "summary": "Capture Pack Payment",
                "parameters": [
                    {
                        "description": "Capture Payment Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.BuyPackCapturePaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/v1/buy-pack/create-order": {
            "post": {
                "description": "Generate a new order for buy a pack",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BuyPack"
                ],
                "summary": "Create Buy Pack Order",
                "parameters": [
                    {
                        "description": "Create Order Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.BuyPackCreateOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.BuyPackCreateOrderResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/v1/buy-pack/detail": {
            "post": {
                "description": "Get info required for buy the pack",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BuyPack"
                ],
                "summary": "Get Buy Pack Details",
                "parameters": [
                    {
                        "description": "details buy pack dto",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.BuyPackDetailsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PackBuyDetailDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/v1/buy-pack/info": {
            "get": {
                "description": "Get config client data for payment system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BuyPack"
                ],
                "summary": "Get Payment Client Data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
...
        "/v1/chili-bank/banks": {
            "get": {
                "description": "Get Chili Banks",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chili Bank"
                ],
                "summary": "Get Banks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.ApiResponse-array_string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/v1/chili-bank/{modelNickname}/accounts": {
            "get": {
                "description": "Get Model Chili Banks Accounts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chili Bank"
                ],
                "summary": "Get Model Accounts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "model nickname",
                        "name": "modelNickname",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.ApiResponse-array_dto_ChiliBankAccountDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "post": {
                "description": "Insert or update a chili bank account for the model",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chili Bank"
                ],
                "summary": "Save Model Account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "model nickname",
                        "name": "modelNickname",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The chili bank account data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ChiliBankAccountDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.ApiResponse-dto_ChiliBankAccountDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/v1/chili-bank/{modelNickname}/{accountId}": {
            "delete": {
                "description": "Delete a chili bank account of the model",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chili Bank"
                ],
                "summary": "Delete Model Account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "model nickname",
                        "name": "modelNickname",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "chili bank account id",
                        "name": "accountId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The chili bank account data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ChiliBankAccountDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.ApiResponse-string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
...
        "/v1/pack/{modelNickName}/{packNumber}/info": {
            "get": {
                "description": "Get information about the pack",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pack"
                ],
                "summary": "Get Pack Info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "model nickname",
                        "name": "modelNickName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "pack number",
                        "name": "packNumber",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.ApiResponse-array_dto_PackInfoDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ChiliBankAccountDTO": {
            "type": "object",
            "required": [
                "accountNumber",
                "accountType",
                "bankName",
                "holderName",
                "rut"
            ],
            "properties": {
                "accountNumber": {
                    "type": "integer",
                    "maximum": 30
                },
                "accountType": {
                    "type": "string",
                    "maxLength": 50
                },
                "bankName": {
                    "type": "string",
                    "maxLength": 255
                },
                "holderName": {
                    "type": "string",
                    "maxLength": 255
                },
                "id": {
                    "type": "string",
                    "maxLength": 255
                },
                "rut": {
                    "type": "integer",
                    "maximum": 10
                }
            }
        },
...
        "dto.PackBuyDetailDto": {
            "type": "object",
            "properties": {
                "modelNickName": {
                    "type": "string"
                },
                "packDollarValue": {
                    "type": "number"
                },
                "packTitle": {
                    "type": "string"
                }
            }
        },
...
        "dto.PackInfoDto": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "handler.BuyPackCapturePaymentRequest": {
            "type": "object",
            "properties": {
                "orderId": {
                    "type": "string"
                }
            }
        },
        "handler.BuyPackCreateOrderRequest": {
            "type": "object",
            "properties": {
                "modelNickName": {
                    "type": "string"
                },
                "packNumber": {
                    "type": "integer"
                },
                "personId": {
                    "type": "string"
                }
            }
        },
        "handler.BuyPackCreateOrderResponse": {
            "type": "object",
            "properties": {
                "orderId": {
                    "type": "string"
                }
            }
        },
        "handler.BuyPackDetailsRequest": {
            "type": "object",
            "properties": {
                "modelNickName": {
                    "type": "string"
                },
                "packNumber": {
                    "type": "integer"
                }
            }
        },
...
        "rest.ApiErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "rest.ApiResponse-array_dto_ChiliBankAccountDTO": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/rest.ApiErrorDetail"
                },
                "payload": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ChiliBankAccountDTO"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "rest.ApiResponse-array_dto_PackInfoDto": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/rest.ApiErrorDetail"
                },
                "payload": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.PackInfoDto"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "rest.ApiResponse-array_string": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/rest.ApiErrorDetail"
                },
                "payload": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "rest.ApiResponse-dto_ChiliBankAccountDTO": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/rest.ApiErrorDetail"
                },
                "payload": {
                    "$ref": "#/definitions/dto.ChiliBankAccountDTO"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "rest.ApiResponse-string": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/rest.ApiErrorDetail"
                },
                "payload": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}